(fix)repository structure & url naming
