+++ config/urls.py
@@ -21,7 +21,6 @@ from accounts.views import duplicate_check
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('api/', include('accounts.urls')),
-    path('api/', include('members.urls')),
     path('api/', include('posts.urls')),
     path('api/', include('notifications.urls')),
     path('api/duplicate', duplicate_check),
