--- members/serializers.py
@@ -1,14 +1,0 @@
-from rest_framework import serializers
-
-from members.models import Member
-
-
-class MemberSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Member
-        fields = (
-            'id',
-            'post',
-            'member',
-            'member_type',
-        )
