+++ posts/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
 
-from posts.models import Post, Comment
+from posts.models import Post, Comment, Member
 
-admin.site.register([Post,Comment])
+admin.site.register([Post, Comment, Member])
