+++ posts/urls.py
@@ -1,9 +1,10 @@
 from rest_framework.routers import DefaultRouter
 
-from .views import PostViewSet, CommentViewSet, PostCommentReadViewSet
+from .views import PostViewSet, CommentViewSet, PostCommentReadViewSet, MemberViewSet
 
 router = DefaultRouter(trailing_slash=False)
-router.register(r'post', PostViewSet, basename='post')
-router.register(r'comment', CommentViewSet, basename='comment')
-router.register(r'post/(?P<postid>[0-9]+)/comment', PostCommentReadViewSet, basename='comment-read')
+router.register(r'posts', PostViewSet, basename='post')
+router.register(r'comments', CommentViewSet, basename='comment')
+router.register(r'posts/(?P<postid>[0-9]+)/comments', PostCommentReadViewSet, basename='comment-read')
+router.register(r'members', MemberViewSet, basename='member')
 urlpatterns = router.urls
